requirements.txt
pandas
yfinance
tabulate
plotly
python-dotenv
